requirements.txt
absl-py==2.3.1
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
bert-score==0.3.13
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chonkie==1.4.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.12.0
fastapi==0.119.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
model2vec==0.7.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nltk==3.8.1
numpy==2.3.4
openai==2.3.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
postgrest==2.22.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
realtime==2.22.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rouge_score==0.1.2
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.22.0
 hints & utilities
typing_extensions==4.15.0
